+++ examples/process_fridge.ex
@@ -6,14 +6,14 @@
 
 object Fridge
   % Creates a new fridge by receiving a list.
-  def constructor(list)
+  def initialize(list)
     % Spawn a new process by invoking the 'loop method defined below.
     % Notice that all methods defined in Fridge are actually defined
     % in Fridge::Proto, this is why we pass Fridge::Proto below.
     pid = Process.spawn Fridge::Proto, 'loop, [list]
 
-    % The values returned by constructor are kept as instance variables.
-    { 'pid: pid }
+    % Returns the object with the pid as instance variable.
+    @('pid: pid)
   end
 
   % Store the given food in the fridge.