+++ src/elixir.erl
@@ -26,16 +26,11 @@ config_change(_Changed, _New, _Remove) ->
 
 % Start elixir as an application.
 start_app() ->
-  % Ensure elixir_object_methods is loaded and running
-  code:ensure_loaded(elixir_object_methods),
+  case lists:keyfind(?MODULE,1, application:loaded_applications()) of
+    false -> application:start(?MODULE);
+    _ -> ok
+  end.
 
-  % Load stdlib files
-  Regexp = cache_regexp(),
-  BasePath = stdlib_path(),
-  BaseFiles = [internal_require(filename:join(BasePath, File), Regexp) || File <- stdlib_files()],
-
-  % Boot the code server with supervisor
-  elixir_sup:start_link([BasePath, BaseFiles]).
 % Boot and process given options. Invoked by Elixir's script.
 start() ->
   start_app(),